Consolidate embedding model test cases into a single test structure (#600)
